+++ src/main/java/com/alchitry/labs/parsers/lucidv2/values/SimpleValueBuilders.kt
@@ -135,7 +135,7 @@ fun BitListValue(
 
 fun BitListValue(value: Long, width: Int, constant: Boolean, signed: Boolean): BitListValue =
     BitListValue(width, constant, signed) {
-        if ((value and (1 shl it).toLong()) != 0.toLong()) Bit.B1 else Bit.B0
+        if ((value and (1L shl it)) != 0L) Bit.B1 else Bit.B0
     }
 
 fun BitListValue(value: Int, width: Int, constant: Boolean, signed: Boolean): BitListValue =
